Fixed alot of stuff

+++ backend/api/Controllers/DepartmentCourseController.cs
@@ -15,15 +15,17 @@ namespace api.Controllers
         private readonly IDepartmentRepository _departmentRepository;
         private readonly ICourseRepository _courseRepository;
         private readonly ICourseDetailsRepository _courseDetailsRepository;
-        public DepartmentCourseController(IDepartmentCourseRepository courseDepRepository, IDepartmentRepository departmentRepository, ICourseRepository courseRepository, ICourseDetailsRepository courseDetailsRepository){
+        private readonly ISemesterDetailsRepository _semesterDetailsRepository;
+        public DepartmentCourseController(IDepartmentCourseRepository courseDepRepository, IDepartmentRepository departmentRepository, ICourseRepository courseRepository, ICourseDetailsRepository courseDetailsRepository, ISemesterDetailsRepository semesterDetailsRepository){
             _departmentCourseRepository = courseDepRepository;
             _departmentRepository = departmentRepository;
             _courseRepository = courseRepository;
             _courseDetailsRepository = courseDetailsRepository;
+            _semesterDetailsRepository = semesterDetailsRepository;
         }
 
         [HttpGet("University/Faculty/Department/Course/")]
-        public async Task<IActionResult> GetDepCourseByCourseIdAndDepId([FromQuery] String DepName, [FromQuery] String CourseName){
+        public async Task<IActionResult> GetDepCourseByCourseAndDepName([FromQuery] String DepName, [FromQuery] String CourseName){
             if(!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -105,11 +107,31 @@ namespace api.Controllers
             if(courseDetails == null){
                 return NotFound("Course Details not found");
             } 
-            
-            var depsDetails = await _departmentCourseRepository.AddCourseToDepAsync(coursePostDto.ToDepartmentCourse());
+
+            var SemesterDetail = await _semesterDetailsRepository.GetSemesterDetailsAsync(coursePostDto.DepartmentName, coursePostDto.TaughtSemester);
+            if(SemesterDetail == null){
+                return NotFound("Semester Details not found. Either that semester doesn't exist or its info is not posted yet.");
+            }
+            int academicYear = (coursePostDto.TaughtSemester + 1) / 2;
+            int totalCourses = await _semesterDetailsRepository.GetNumOfCoursesInAcademicYear(academicYear);
+
+            if(totalCourses == -1){
+                return NotFound("No semester details on the specified academic year.");
+            }
+            int code = academicYear*1000 + totalCourses + 1;
+            String CourseCode = validDepartment.DepCode + "-" + code.ToString();
+
+            var depsDetails = await _departmentCourseRepository.AddCourseToDepAsync(coursePostDto.ToDepartmentCourse(CourseCode));
             
             if(depsDetails == null){
                 return BadRequest();
+            }
+            
+            SemesterDetail.TotalCourses++;
+            var result = await _semesterDetailsRepository.UpdateSemesterDetailsAsync(SemesterDetail);
+
+            if(result == null){
+                return StatusCode(500, "Failed to increment semester courses.");
             }
 
             return Ok(depsDetails.ToDepartmentCourseDto());
@@ -122,18 +144,23 @@ namespace api.Controllers
                 return BadRequest(ModelState);
             }
 
-            var depCourse = await _departmentCourseRepository.GetDeparmentCourseAsync(CourseName, DepName);
-
-            if(depCourse == null){
-                return BadRequest();
-            }
-
             if(departmentCourseUpdateDto.CourseName != CourseName || departmentCourseUpdateDto.DepartmentName != DepName){
                 return BadRequest();
             }
 
             if(departmentCourseUpdateDto.Status != "Open" && departmentCourseUpdateDto.Status != "Closed"){
                 return BadRequest("Status can only be Open or Closed");
+            }
+
+            var SemesterDetail = await _semesterDetailsRepository.GetSemesterDetailsAsync(departmentCourseUpdateDto.DepartmentName, departmentCourseUpdateDto.TaughtSemester);
+            if(SemesterDetail == null){
+                return NotFound("Semester Details not found. Either that semester doesn't exist or its info is not posted yet. Failed to Update the Course.");
+            }
+
+            var depCourse = await _departmentCourseRepository.GetDeparmentCourseAsync(CourseName, DepName);
+
+            if(depCourse == null){
+                return BadRequest();
             }
 
             depCourse.TaughtSemester = departmentCourseUpdateDto.TaughtSemester;
@@ -156,9 +183,29 @@ namespace api.Controllers
                 return BadRequest(ModelState);
             }
 
-            var result = await _departmentCourseRepository.DeleteCourseFromDepAsync(CourseName, DepName);
+            var validCourse = await _departmentCourseRepository.GetDeparmentCourseAsync(CourseName, DepName);
+
+            if(validCourse == null){
+                return NotFound();
+            }
+            
+            var SemesterDetail = await _semesterDetailsRepository.GetSemesterDetailsAsync(validCourse.DepartmentName, validCourse.TaughtSemester);
+            if(SemesterDetail == null){
+                return NotFound("Semester Details not found. Either that semester doesn't exist or its info is not posted yet. Failed to Delete the Course.");
+            }
+            // Should not decrese because CourseCode depends on it 
+            /*
+            SemesterDetail.TotalCourses--;
+            var result = await _semesterDetailsRepository.UpdateSemesterDetailsAsync(SemesterDetail);
 
             if(result == null){
+                return StatusCode(500, "Failed to decrease semester courses. Aborting course deletion.");
+            }
+            */
+
+            var deletionResult = await _departmentCourseRepository.DeleteCourseFromDepAsync(CourseName, DepName);
+
+            if(deletionResult == null){
                 return BadRequest();
             }
